Reverted from obsolete rubygems/open-uri back to open-uri.
This is a revert of commits 0628cc7... and  88e9160... which introduced rubygems/open-uri in order to allow for proxy settings to be used.  However open-uri has honoured ENV['http_proxy'] and cousins for some time.  Moreover, rubygems/open-uri won't be fond in recent rubygems distros, having been removed in Nov. 2006.

Signed-off-by: Chris Wanstrath <chris@ozmm.org>